moved states to constants
